Update keystone.conf to reflect Grizzly changes
This change moves keystone.conf forward to Grizzly. In addition, as
the comments provided in the openstack/keystone.git source are quite
helpful, this change seeks to make this template as consitent as
possible with the sample source provided upstream.

With the development of support for conf.d directories in Oslo, this
file will hopefully be broken out accordingly.